Moved test files from app/ocmponents and app/lib from subfolders in to main folders.
