+++ app/components/Slides/index.test.js
@@ -5,7 +5,7 @@ import { shallow } from 'enzyme';
 
 import { dummyContentItemData as dummyData, dummyProviderProps } from 'lib/testResources';
 
-import { PureSlides } from '..';
+import { PureSlides } from '.';
 
 describe(`Slides`, (): void => {
